requirements.txt
pyOpenSSL>=0.14
dnspython>=1.11.1